Added git ignore. Start of script.
Was attempting to grab the contents of every file in every commit.
Doesn't look like this library supports that.
Right now, it grabs the entire contents of every file that was modified in a commit.
Seems like doing it this way will be more effiecent though, avoids reading in the same text that is never changed multiple times.